refactor(cli): move install.rs tests to sibling install_tests.rs to stay under the per-file line cap (#964)
